requirements.txt
numpy==1.17.4
psycopg2-binary==2.8.2
boto3==1.12.20